Zotero + Api cleanup, dashboard separated
