+++ Angular/src/app/dashboard/fragments-dashboard/fragments-dashboard.component.scss
@@ -1,0 +1,1 @@
+@import url("../dashboard.component.scss");
